+++ R/HDI.R
@@ -1,17 +1,18 @@
 #' @title Compute statistics for MCMC samples
 #' @name hdi
 #'
-#' @description \code{hdi()} computes the high density interval for values from
+#' @description \code{hdi()} computes the highest density interval for values from
 #'   MCMC samples. \code{rope()} calculates the proportion of a posterior
 #'   distribution that lies within a region of practical equivalence.
 #'   \code{n_eff()} calculates the number of effective samples (effective
 #'   sample size). \code{mcse()} returns the Monte Carlo standard error.
 #'
 #' @param x A \code{stanreg}, \code{stanfit}, or \code{brmsfit} object. For
-#'   \code{hdi()} and \code{rope()}, may also be a vector of values from a
-#'   probability distribution (e.g., posterior probabilities from MCMC sampling).
-#' @param prob Scalar between 0 and 1, indicating the mass within the credible
-#'   interval that is to be estimated.
+#'   \code{hdi()} and \code{rope()}, may also be a data frame or a vector
+#'   of values from a probability distribution (e.g., posterior probabilities
+#'   from MCMC sampling).
+#' @param prob Vector of scalars between 0 and 1, indicating the mass within
+#'   the credible interval that is to be estimated. See \code{\link{hdi}}.
 #' @param rope Vector of length two, indicating the lower and upper limit of a
 #'   range around zero, which indicates the region of practical equivalence.
 #'   Values of the posterior distribution within this range are considered as
@@ -30,11 +31,16 @@
 #' @return For \code{hdi()}, if \code{x} is a vector, returns a vector of length
 #'   two with the lower and upper limit of the HDI; if \code{x} is a
 #'   \code{stanreg}, \code{stanfit} or \code{brmsfit} object, returns a
-#'   tibble with lower and upper HDI-limits for each predictor.
-#'   For \code{rope()}, returns the proportion of values from \code{x}
-#'   that are within the boundaries of \code{rope}. \code{mcse()} and
-#'   \code{n_eff()} return a tibble with two columns: one with the term names
-#'   and one with the related statistic.
+#'   tibble with lower and upper HDI-limits for each predictor. To distinguish
+#'   multiple HDI values, column names for the HDI get a suffix when \code{prob}
+#'   has more than one element.
+#'   \cr \cr
+#'   For \code{rope()}, returns a tibble with two columns: the proportion of
+#'   values from \code{x} that are within and outside the boundaries of
+#'   \code{rope}.
+#'   \cr \cr
+#'   \code{mcse()} and \code{n_eff()} return a tibble with two columns: one
+#'   with the term names and one with the related statistic.
 #'
 #' @details Computation for HDI is based on the code from Kruschke 2015, pp. 727f.
 #'   For default sampling in Stan (4000 samples), the 90\% intervals for HDI are
@@ -54,6 +60,9 @@
 #' if (require("rstanarm")) {
 #'   fit <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
 #'   hdi(fit)
+#'
+#'   # return multiple intervals
+#'   hdi(fit, prob = c(.5, .7, .9))
 #'
 #'   # fit logistic regression model
 #'   fit <- stan_glm(
@@ -75,9 +84,6 @@
 #'   rope(fit, rope = c(.8, 1.25), trans = exp)
 #' }}
 #'
-#' @importFrom tibble as_tibble rownames_to_column
-#' @importFrom purrr map_dbl map_df
-#' @importFrom sjmisc rotate_df
 #' @export
 hdi <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", "all")) {
   UseMethod("hdi")
@@ -89,14 +95,7 @@ hdi.stanreg <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", 
   # check arguments
   type <- match.arg(type)
 
-  # get posterior data
-  dat <- x %>%
-    tibble::as_tibble() %>%
-    purrr::map_df(~ hdi_helper(.x, prob, trans)) %>%
-    sjmisc::rotate_df() %>%
-    tibble::rownames_to_column()
-
-  colnames(dat) <- c("term", "hdi.low", "hdi.high")
+  dat <- hdi_worker(x = x, prob = prob, trans = trans)
 
   # check if we need to remove random or fixed effects
   remove_effects_from_stan(dat, type, is.brms = FALSE)
@@ -112,14 +111,7 @@ hdi.brmsfit <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", 
   if (!requireNamespace("brms", quietly = TRUE))
     stop("Please install and load package `brms` first.")
 
-  # get posterior data
-  dat <- x %>%
-    tibble::as_tibble() %>%
-    purrr::map_df(~ hdi_helper(.x, prob, trans)) %>%
-    sjmisc::rotate_df() %>%
-    tibble::rownames_to_column()
-
-  colnames(dat) <- c("term", "hdi.low", "hdi.high")
+  dat <- hdi_worker(x = x, prob = prob, trans = trans)
 
   # check if we need to remove random or fixed effects
   remove_effects_from_stan(dat, type, is.brms = TRUE)
@@ -131,17 +123,19 @@ hdi.stanfit <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", 
   # check arguments
   type <- match.arg(type)
 
-  # get posterior data
-  dat <- x %>%
-    as.data.frame() %>%
-    purrr::map_df(~ hdi_helper(.x, prob, trans)) %>%
-    sjmisc::rotate_df() %>%
-    tibble::rownames_to_column()
-
-  colnames(dat) <- c("term", "hdi.low", "hdi.high")
+  dat <- hdi_worker(x = x, prob = prob, trans = trans)
 
   # check if we need to remove random or fixed effects
   remove_effects_from_stan(dat, type, is.brms = FALSE)
+}
+
+
+#' @export
+hdi.data.frame <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", "all")) {
+  # check arguments
+  type <- match.arg(type)
+
+  hdi_worker(x = x, prob = prob, trans = trans)
 }
 
 
@@ -151,7 +145,50 @@ hdi.default <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", 
 }
 
 
+#' @importFrom tibble as_tibble rownames_to_column
+#' @importFrom purrr map_df
+#' @importFrom sjmisc rotate_df
+hdi_worker <- function(x, prob, trans) {
+  dat <- purrr::map(
+    prob,
+    function(i) {
+
+      out <- x %>%
+        tibble::as_tibble() %>%
+        purrr::map_df(~ hdi_helper(.x, i, trans)) %>%
+        sjmisc::rotate_df() %>%
+        tibble::rownames_to_column()
+
+      colnames(out) <- c("term", "hdi.low", "hdi.high")
+      out
+
+    }) %>%
+    dplyr::bind_cols() %>%
+    dplyr::select(1, tidyselect::starts_with("hdi."))
+
+
+
+  # for multiple HDIs, fix column names
+
+  if (length(prob) > 1) {
+    suffix <- prob %>%
+      purrr::map(~ rep(.x, 2)) %>%
+      purrr::flatten_dbl()
+
+    colnames(dat)[2:ncol(dat)] <-
+      sprintf(
+        "%s_%s",
+        rep(c("hdi.low", "hdi.high"), length(prob)),
+        as.character(suffix)
+      )
+  }
+
+  dat
+}
+
+
 # based on Kruschke 2015, pp727f
+#' @importFrom purrr map_dbl map_df
 hdi_helper <- function(x, prob, trans) {
   x <- sort(x)
   ci.index <- ceiling(prob * length(x))