+++ R/eta_sq.R
@@ -1,16 +1,33 @@
 #' @title Effect size statistics for anova
 #' @name eta_sq
-#' @description Returns the (partial) eta-squared, omega-squared statistic
-#'              or Cohen's F for all terms in an anovas. \code{anova_stats()} returns
-#'              a tidy summary, including all these statistics and power for each term.
+#' @description Returns the (partial) eta-squared, (partial) omega-squared statistic
+#'   or Cohen's F for all terms in an anovas. \code{anova_stats()} returns
+#'   a tidy summary, including all these statistics and power for each term.
 #'
 #' @param model A fitted anova-model of class \code{aov} or \code{anova}. Other
-#'              models are coerced to \code{\link[stats]{anova}}.
+#'   models are coerced to \code{\link[stats]{anova}}.
 #' @param partial Logical, if \code{TRUE}, the partial eta-squared is returned.
 #' @param digits Number of decimal points in the returned data frame.
+#' @param ci.lvl Scalar between 0 and 1. If not \code{NULL}, returns a data
+#'   frame with effect sizes including lower and upper confidence intervals.
 #'
-#' @return A numeric vector with the effect size statistics; for \code{anova_stats()},
-#'         a tidy data frame with all these statistics.
+#' @inheritParams bootstrap
+#'
+#' @return A data frame with the term name(s) and effect size statistics; if
+#'   \code{ci.lvl} is not \code{NULL}, a data frame including lower and
+#'   upper confidence intervals is returned. For \code{anova_stats()}, a tidy
+#'   data frame with all statistics is returned (excluding confidence intervals).
+#'
+#' @details For \code{eta_sq()} (with \code{partial = FALSE}), due to
+#'   non-symmetry, confidence intervals are based on bootstrap-methods. In this
+#'   case, \code{n} indicates the number of bootstrap samples to be drawn to
+#'   compute the confidence intervals.
+#'   \cr \cr
+#'   For partial eta-squared (\code{eta_sq()} with \code{partial = TRUE}),
+#'   confidence intervals are based on \code{\link[apaTables]{get.ci.partial.eta.squared}}
+#'   and for omega-squared, confidence intervals are based on
+#'   \code{\link[MBESS]{conf.limits.ncf}}. Confidence intervals for partial
+#'   omega-squared is also based on bootstrapping.
 #'
 #' @references Levine TR, Hullett CR (2002): Eta Squared, Partial Eta Squared, and Misreporting of Effect Size in Communication Research (\href{https://www.msu.edu/~levinet/eta\%20squared\%20hcr.pdf}{pdf})
 #'
@@ -28,31 +45,113 @@
 #' omega_sq(fit)
 #' eta_sq(fit, partial = TRUE)
 #'
+#' # CI for omega squared requires MBESS packages
+#' if (requireNamespace("MBESS", quietly = TRUE)) {
+#'   omega_sq(fit, ci.lvl = .95)
+#' }
+#'
 #' anova_stats(car::Anova(fit, type = 2))
 #'
-#' @importFrom broom tidy
-#' @importFrom purrr map_dbl
-#' @importFrom stats anova
+#' @importFrom tibble tibble
 #' @export
-eta_sq <- function(model, partial = FALSE) {
+eta_sq <- function(model, partial = FALSE, ci.lvl = NULL, n = 1000) {
+
   if (partial)
-    aov_stat(model, type = "peta")
+    type <- "peta"
   else
-    aov_stat(model, type = "eta")
+    type <- "eta"
+
+
+  es <- aov_stat(model, type = type)
+
+  x <- tibble::tibble(
+    term = names(es),
+    es = es
+  )
+
+
+  if (partial) {
+    if (!is.null(ci.lvl) && !is.na(ci.lvl)) {
+      x <- dplyr::bind_cols(x, peta_sq_ci(aov.sum = aov_stat_summary(model), ci.lvl = ci.lvl))
+    }
+  } else {
+    if (!is.null(ci.lvl) && !is.na(ci.lvl)) {
+      x <-
+        es_boot_fun(
+          model = model,
+          type = "eta",
+          ci.lvl = ci.lvl,
+          n = n
+        )
+    }
+  }
+
+  colnames(x)[2] <- dplyr::case_when(
+    type == "eta" ~ "etasq",
+    type == "peta" ~ "partial.etasq",
+    TRUE ~ "effect.size"
+  )
+
+  x
 }
 
 
+
 #' @rdname eta_sq
+#' @importFrom dplyr bind_cols mutate
+#' @importFrom tibble tibble
 #' @export
-omega_sq <- function(model) {
-  aov_stat(model, type = "omega")
+omega_sq <- function(model, partial = FALSE, ci.lvl = NULL, n = 1000) {
+
+  if (partial)
+    type <- "pomega"
+  else
+    type <- "omega"
+
+
+  es <- aov_stat(model, type = type)
+
+  x <- tibble::tibble(
+    term = names(es),
+    es = es
+  )
+
+
+  if (partial) {
+    if (!is.null(ci.lvl) && !is.na(ci.lvl)) {
+      x <-
+      es_boot_fun(
+        model = model,
+        type = "pomega",
+        ci.lvl = ci.lvl,
+        n = n
+      )
+    }
+  } else {
+    if (!is.null(ci.lvl) && !is.na(ci.lvl)) {
+      x <- dplyr::bind_cols(x, omega_sq_ci(aov.sum = aov_stat_summary(model), ci.lvl = ci.lvl))
+    }
+  }
+
+  colnames(x)[2] <- dplyr::case_when(
+    type == "omega" ~ "omegasq",
+    type == "pomega" ~ "partial.omegasq",
+    TRUE ~ "effect.size"
+  )
+
+  x
 }
 
 
 #' @rdname eta_sq
 #' @export
 cohens_f <- function(model) {
-  aov_stat(model, type = "cohens.f")
+  es <- aov_stat(model, type = "cohens.f")
+
+  tibble::tibble(
+    term = names(es),
+    cohens.f = es
+  )
 }
 
 
@@ -72,13 +171,14 @@ anova_stats <- function(model, digits = 3) {
   etasq <- aov_stat_core(aov.sum, type = "eta")
   partial.etasq <- aov_stat_core(aov.sum, type = "peta")
   omegasq <- aov_stat_core(aov.sum, type = "omega")
+  partial.omegasq <- aov_stat_core(aov.sum, type = "pomega")
 
   # compute power for each estimate
   cohens.f <- sqrt(partial.etasq / (1 - partial.etasq))
 
   # bind as data frame
-  as <- tibble::tibble(etasq, partial.etasq, omegasq, cohens.f) %>%
-    tibble::add_row(etasq = NA, partial.etasq = NA, omegasq = NA, cohens.f = NA) %>%
+  as <- tibble::tibble(etasq, partial.etasq, omegasq, partial.omegasq, cohens.f) %>%
+    tibble::add_row(etasq = NA, partial.etasq = NA, omegasq = NA, partial.omegasq = NA, cohens.f = NA) %>%
     sjmisc::add_columns(aov.sum)
 
   # get nr of terms
@@ -110,10 +210,14 @@ aov_stat <- function(model, type) {
 aov_stat_summary <- function(model) {
   # check that model inherits from correct class
   # else, try to coerce to anova table
-  if (!inherits(model, c("aov", "anova"))) model <- stats::anova(model)
+  if (!inherits(model, c("aov", "anova", "anova.rms"))) model <- stats::anova(model)
 
   # get summary table
   aov.sum <- broom::tidy(model)
+
+  # need special handling for rms-anova
+  if (inherits(model, "anova.rms"))
+    colnames(aov.sum) <- c("term", "df", "sumsq", "meansq", "statistic", "p.value")
 
   # for car::Anova, the meansq-column might be missing, so add it manually
   if (!tibble::has_name(aov.sum, "meansq"))
@@ -135,12 +239,23 @@ aov_stat_core <- function(aov.sum, type) {
   # number of terms in model
   n_terms <- nrow(aov.sum) - 1
 
+  # number of observations
+  N <- sum(aov.sum[["df"]]) + 1
+
+
   if (type == "omega") {
     # compute omega squared for each model term
     aovstat <- purrr::map_dbl(1:n_terms, function(x) {
       ss.term <- aov.sum[["sumsq"]][x]
       df.term <- aov.sum[["df"]][x]
       (ss.term - df.term * meansq.resid) / (ss.total + meansq.resid)
+    })
+  } else if (type == "pomega") {
+    # compute partial omega squared for each model term
+    aovstat <- purrr::map_dbl(1:n_terms, function(x) {
+      df.term <- aov.sum[["df"]][x]
+      meansq.term <- aov.sum[["meansq"]][x]
+      (df.term * (meansq.term - meansq.resid)) / (df.term * meansq.term + (N - df.term) * meansq.resid)
     })
   } else if (type == "eta") {
     # compute eta squared for each model term
@@ -159,4 +274,122 @@ aov_stat_core <- function(aov.sum, type) {
   names(aovstat) <- aov.sum[["term"]][1:n_terms]
 
   aovstat
+}
+
+
+#' @importFrom tibble tibble
+#' @importFrom purrr map_df
+omega_sq_ci <- function(aov.sum, ci.lvl = .95) {
+
+  if (!requireNamespace("MBESS", quietly = TRUE))
+    stop("Package `MBESS` needed to compute confidence intervals. Pleas install that package first.")
+
+  rows <- nrow(aov.sum) - 1
+  df.den <- aov.sum[["df"]][rows + 1]
+  N <- sum(aov.sum[["df"]]) + 1
+
+  purrr::map_df(
+    1:rows,
+    function(.x) {
+      df.num = aov.sum[.x, "df"]
+
+      ci <- MBESS::conf.limits.ncf(
+        F.value = aov.sum[.x, "statistic"],
+        conf.level = ci.lvl,
+        df.1 = df.num,
+        df.2 = df.den
+      )
+
+      ci.low <- ci$Lower.Limit / (ci$Lower.Limit + N)
+      ci.high <- ci$Upper.Limit / (ci$Upper.Limit + N)
+
+      tibble::tibble(
+        conf.low = ci.low,
+        conf.high = ci.high
+      )
+    }
+  )
+}
+
+
+#' @importFrom tibble tibble
+#' @importFrom purrr map_df
+peta_sq_ci <- function(aov.sum, ci.lvl = .95) {
+
+  if (!requireNamespace("apaTables", quietly = TRUE))
+    stop("Package `apaTables` needed to compute confidence intervals. Pleas install that package first.")
+
+  rows <- nrow(aov.sum) - 1
+  df.den <- aov.sum[["df"]][rows + 1]
+
+  purrr::map_df(
+    1:rows,
+    function(.x) {
+      df.num = aov.sum[.x, "df"]
+
+      ci <- apaTables::get.ci.partial.eta.squared(
+        F.value = aov.sum[.x, "statistic"],
+        df1 = df.num,
+        df2 = df.den,
+        conf.level = ci.lvl
+      )
+
+      tibble::tibble(
+        conf.low = ci$LL,
+        conf.high = ci$UL
+      )
+    }
+  )
+}
+
+
+#' @importFrom broom tidy
+#' @importFrom purrr map map_df
+#' @importFrom dplyr bind_cols mutate case_when pull
+#' @importFrom tibble tibble
+#' @importFrom stats anova formula
+#' @importFrom sjmisc rotate_df
+es_boot_fun <- function(model, type, ci.lvl, n) {
+
+  es <- aov_stat(model = model, type = type)
+
+  x <- tibble::tibble(
+    term = names(es),
+    es = es
+  )
+
+  mdata <- sjstats::model_frame(model)
+  mformula <- stats::formula(model)
+
+  # this is a bit sloppy, but I need to catch all exceptions here
+  # if we have a 1-way-anova, map() could return a column with
+  # one value per row (a vector). However, if the model has more
+  # covariates/factors, map() returns a list-colum with 3 values
+  # per row, which need to be spread into a 3 columns data frame.
+
+  es <- mdata %>%
+    bootstrap(n = n) %>%
+    dplyr::mutate(es = purrr::map(
+      .data$strap,
+      function(i) {
+        m <- lm(mformula, data = i)
+        dat <- aov_stat(m, type = type)
+        sjmisc::rotate_df(as.data.frame(dat))
+      }
+    )) %>%
+    dplyr::pull(2) %>%
+    purrr::map_df(~ .x) %>%
+    boot_ci()
+
+  x <- dplyr::bind_cols(x, es[, -1])
+
+  colnames(x)[2] <- dplyr::case_when(
+    type == "eta" ~ "etasq",
+    type == "peta" ~ "partial.etasq",
+    type == "omega" ~ "omegasq",
+    type == "pomega" ~ "partial.omegasq",
+    TRUE ~ "effect.size"
+  )
+
+  x
 }