+++ R/icc.R
@@ -1,25 +1,34 @@
 #' @title Intraclass-Correlation Coefficient
 #' @name icc
 #' @description This function calculates the intraclass-correlation
-#'                (icc) - sometimes also called \emph{variance partition coefficient}
-#'                (vpc) - for random intercepts of mixed effects models.
-#'                Currently, \code{\link[lme4]{merMod}}, \code{\link[glmmTMB]{glmmTMB}}
-#'                \code{stanreg} and \code{\link[brms]{brmsfit}}
-#'                objects are supported.
+#'    (icc) - sometimes also called \emph{variance partition coefficient}
+#'    (vpc) - for random intercepts of mixed effects models. Currently,
+#'    \code{\link[lme4]{merMod}}, \code{\link[glmmTMB]{glmmTMB}},
+#'    \code{stanreg} and \code{\link[brms]{brmsfit}} objects are supported.
 #'
 #' @param x Fitted mixed effects model (of class \code{merMod}, \code{glmmTMB},
-#'          \code{stanreg} or \code{brmsfit}).
+#'    \code{stanreg} or \code{brmsfit}).
 #' @param ... More fitted model objects, to compute multiple intraclass-correlation
-#'              coefficients at once.
+#'    coefficients at once.
+#' @param posterior Logical, if \code{TRUE} and \code{x} is a \code{brmsfit}
+#'    object, ICC values are computed for each sample of the posterior
+#'    distribution. In this case, a data frame is returned with the same
+#'    number of rows as samples in \code{x}, with one column per random
+#'    effect ICC.
 #'
-#' @return A numeric vector with all random intercept intraclass-correlation-coefficients,
-#'           or a list of numeric vectors, when more than one model were used
-#'           as arguments. Furthermore, between- and within-group variances as well
-#'           as random-slope variance are returned as attributes.
+#' @return If \code{posterior = FALSE} (the default), a numeric vector with all
+#'    random intercept intraclass-correlation-coefficients, or a list of
+#'    numeric vectors, when more than one model were used as arguments.
+#'    Furthermore, between- and within-group variances as well as random-slope
+#'    variance are returned as attributes.
+#'    \cr \cr
+#'    If \code{posterior = TRUE}, \code{icc()} returns a data frame with ICC
+#'    and variance components for each sample of the posterior distribution.
 #'
 #' @references \itemize{
 #'               \item Aguinis H, Gottfredson RK, Culpepper SA. 2013. Best-Practice Recommendations for Estimating Cross-Level Interaction Effects Using Multilevel Modeling. Journal of Management 39(6): 1490–1528 (\doi{10.1177/0149206313478188})
 #'               \item Aly SS, Zhao J, Li B, Jiang J. 2014. Reliability of environmental sampling culture results using the negative binomial intraclass correlation coefficient. Springerplus 14(3) (\doi{10.1186/2193-1801-3-40})
+#'               \item Goldstein H, Browne W, Rasbash J. 2010. Partitioning Variation in Multilevel Models. Understanding Statistics, 1:4, 223-231 (\doi{10.1207/S15328031US0104_02})
 #'               \item Grace-Martion K. The Intraclass Correlation Coefficient in Mixed Models, \href{http://www.theanalysisfactor.com/the-intraclass-correlation-coefficient-in-mixed-models/}{web}
 #'               \item Hox J. 2002. Multilevel analysis: techniques and applications. Mahwah, NJ: Erlbaum
 #'               \item Rabe-Hesketh S, Skrondal A. 2012. Multilevel and longitudinal modeling using Stata. 3rd ed. College Station, Tex: Stata Press Publication
@@ -70,12 +79,26 @@
 #'          intercept variance) by the total variance (i.e. sum of between-group-variance
 #'          and within-group (residual) variance). \cr \cr
 #'       The calculation of the ICC for generalized linear mixed models with binary outcome is based on
-#'       \cite{Wu et al. (2012)}. For Poisson multilevel models, please refere to \cite{Stryhn et al. (2006)}.
+#'       \cite{Wu et al. (2012)}. For Poisson multilevel models, please refer to \cite{Stryhn et al. (2006)}.
 #'       \cite{Aly et al. (2014)} describe computation of ICC for negative binomial models.
 #'       \cr \cr
-#'       There is a \code{print}-method that prints the variance parameters using
-#'       the \code{comp}-argument set to \code{"var"}: \code{print(x, comp = "var")}
-#'       (see 'Examples'). The \code{\link{re_var}}-function is a convenient wrapper.
+#'       \strong{Caution:} For models with random slopes and random intercepts,
+#'       the ICC would differ at each unit of the predictors. Hence, the ICC for these
+#'       kind of models cannot be understood simply as proportion of variance
+#'       (see \cite{Goldstein et al. 2010}). For convenience reasons, as the
+#'       \code{icc()} function also extracts the different random effects
+#'       variances, the ICC for random-slope-intercept-models is reported
+#'       nonetheless, but it is usually no meaningful summary of the
+#'       proportion of variances.
+#'       \cr \cr
+#'       If \code{posterior = FALSE}, there is a \code{print()}-method that prints
+#'       the variance parameters using the \code{comp}-argument set to \code{"var"}:
+#'       \code{print(x, comp = "var")} (see 'Examples'). The
+#'       \code{\link{re_var}}-function is a convenient wrapper. If
+#'       \code{posterior = TRUE}, the \code{print()}-method accepts the arguments
+#'       \code{prob} and \code{digits}, which indicate the probability of the
+#'       uncertainty interval for the ICC and variance components, and the digits
+#'       in the output (see also 'Examples').
 #'       \cr \cr
 #'       The random effect variances indicate the between- and within-group
 #'         variances as well as random-slope variance and random-slope-intercept
@@ -95,11 +118,13 @@
 #' fit0 <- lmer(Reaction ~ 1 + (1 | Subject), sleepstudy)
 #' icc(fit0)
 #'
+#' # note: ICC for random-slope-intercept model usually not
+#' # meaningful - see 'Note'.
 #' fit1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
 #' icc(fit1)
 #'
 #' sleepstudy$mygrp <- sample(1:45, size = 180, replace = TRUE)
-#' fit2 <- lmer(Reaction ~ Days + (1 | mygrp) + (Days | Subject), sleepstudy)
+#' fit2 <- lmer(Reaction ~ Days + (1 | mygrp) + (1 | Subject), sleepstudy)
 #' icc(fit2)
 #'
 #' # return icc for all models at once
@@ -111,11 +136,45 @@
 #' print(icc1, comp = "var")
 #' print(icc2, comp = "var")
 #'
+#' \dontrun{
+#' # compute ICC for Bayesian mixed model, with an ICC for each
+#' # sample of the posterior. The print()-method then shows
+#' # the median ICC as well as 89% HDI for the ICC.
+#' # Change interval with print-method:
+#' # print(icc(m, posterior = TRUE), prob = .5)
+#'
+#' if (requireNamespace("brms", quietly = TRUE)) {
+#'   library(dplyr)
+#'   sleepstudy$mygrp <- sample(1:5, size = 180, replace = TRUE)
+#'   sleepstudy <- sleepstudy %>%
+#'     group_by(mygrp) %>%
+#'     mutate(mysubgrp = sample(1:30, size = n(), replace = TRUE))
+#'   m <- brm(
+#'     Reaction ~ Days + (1 | mygrp / mysubgrp) + (1 | Subject),
+#'     data = sleepstudy
+#'   )
+#'
+#'   # by default, 89% interval
+#'   icc(m, posterior = TRUE)
+#'
+#'   # show 50% interval
+#'   print(icc(m, posterior = TRUE), prob = .5, digits = 2)
+#' }}
 #'
 #' @importFrom purrr map2
 #' @export
-icc <- function(x, ...) {
+icc <- function(x, ..., posterior = FALSE) {
+
+  if (isTRUE(posterior) && !inherits(x, "brmsfit")) {
+    warning("ICC from posterior samples only possible for `brmsfit`-objects.")
+    posterior <- FALSE
+  }
+
   # return value
+  if (posterior) {
+    return(icc.posterior(x, deparse(substitute(x))))
+  }
+
   icc_ <- icc.lme4(x, deparse(substitute(x)))
 
   # check if we have multiple parameters
@@ -160,6 +219,7 @@ icc <- function(x, ...) {
 #' @importFrom stats family formula
 #' @importFrom purrr map map_dbl map_lgl
 #' @importFrom sjmisc str_contains
+#' @importFrom tibble has_name
 icc.lme4 <- function(fit, obj.name) {
   # check object class
   if (is_merMod(fit) || inherits(fit, c("glmmTMB", "brmsfit"))) {
@@ -176,7 +236,7 @@ icc.lme4 <- function(fit, obj.name) {
     is_negbin <-
       sjmisc::str_contains(
         fitfam,
-        c("Negative Binomial", "nbinom"),
+        c("Negative Binomial", "nbinom", "negbinomial"),
         ignore.case = TRUE,
         logic = "OR"
       )
@@ -198,23 +258,23 @@ icc.lme4 <- function(fit, obj.name) {
     if (inherits(fit, "glmmTMB")) {
       reva <- glmmTMB::VarCorr(fit)[[1]]
     } else if (inherits(fit, "brmsfit")) {
-      reva <- brms::VarCorr(fit, old = TRUE)
+      reva <- brms::VarCorr(fit)
     } else
       reva <- lme4::VarCorr(fit)
 
 
-    # for brmsfit-objects, remove "RESIDUAL" element from list
+    # for brmsfit-objects, remove "residual__" element from list
     # and save in separate object
     if (inherits(fit, "brmsfit")) {
-      reva.resid <- reva[names(reva) == "RESIDUAL"]
-      reva <- reva[!(names(reva) == "RESIDUAL")]
+      reva.resid <- reva[names(reva) == "residual__"]
+      reva <- reva[!(names(reva) == "residual__")]
     }
 
 
     # retrieve only intercepts
 
     if (inherits(fit, "brmsfit"))
-      vars <- purrr::map(reva, ~ .x$cov$mean[1])
+      vars <- purrr::map(reva, ~ .x$sd[1] ^ 2)
     else
       vars <- purrr::map(reva, ~ .x[1])
 
@@ -227,14 +287,14 @@ icc.lme4 <- function(fit, obj.name) {
 
     # random slope-variances (tau 11)
     if (inherits(fit, "brmsfit"))
-      tau.11 <- unlist(lapply(reva, function(x) diag(x$cov$mean)[-1]))
+      tau.11 <- unlist(lapply(reva, function(x) diag(x$cov[, 1, ])[-1]))
     else
       tau.11 <- unlist(lapply(reva, function(x) diag(x)[-1]))
 
 
     # get residual standard deviation sigma
     if (inherits(fit, "brmsfit"))
-      sig <- reva.resid$RESIDUAL$sd[1]
+      sig <- reva.resid[["residual__"]]$sd[1]
     else
       sig <- attr(reva, "sc")
 
@@ -256,7 +316,7 @@ icc.lme4 <- function(fit, obj.name) {
       # for logistic models, we use pi / 3
       resid_var <- (pi ^ 2) / 3
     } else if (inherits(fit, "glmerMod") && is_negbin) {
-      # for negative binomial models, we use 0
+      # for negative binomial models, we use 1
       resid_var <- 1
     } else {
       # for linear and poisson models, we have a clear residual variance
@@ -276,6 +336,11 @@ icc.lme4 <- function(fit, obj.name) {
         beta <- as.numeric(lme4::fixef(fit)["(Intercept)"])
         # ... and the theta value to compute the ICC
         r <- lme4::getME(fit, "glmer.nb.theta")
+      } else if (inherits(fit, "brms")) {
+        # for negative binomial models, we also need the intercept...
+        beta <- as.numeric(brms::fixef(fit)[[1]])
+        # ... and the theta value to compute the ICC
+        r <- sig
       } else {
         # for negative binomial models, we also need the intercept...
         beta <- as.numeric(glmmTMB::fixef(fit)[[1]]["(Intercept)"])
@@ -300,7 +365,7 @@ icc.lme4 <- function(fit, obj.name) {
     # do we have any rnd slopes?
 
     if (inherits(fit, "brmsfit"))
-      has_rnd_slope <- purrr::map_lgl(reva, ~ dim(.x$cor$mean)[1] > 1)
+      has_rnd_slope <- purrr::map_lgl(reva, ~ tibble::has_name(.x, "cor"))
     else
       has_rnd_slope <- purrr::map_lgl(reva, ~ dim(attr(.x, "correlation"))[1] > 1)
 
@@ -315,9 +380,9 @@ icc.lme4 <- function(fit, obj.name) {
 
       if (inherits(fit, "brmsfit")) {
         # get slope-intercept-correlations
-        rho.01 <- purrr::map_dbl(rnd_slope, ~ .x$cor$mean[1, 2])
+        rho.01 <- purrr::map_dbl(rnd_slope, ~ .x$cor[1, 1, 2])
         # get standard deviations, multiplied
-        std_ <- purrr::map_dbl(rnd_slope, ~ prod(.x$sd))
+        std_ <- purrr::map_dbl(rnd_slope, ~ prod(.x$sd[, 1]))
       } else {
         # get slope-intercept-correlations
         rho.01 <- purrr::map_dbl(rnd_slope, ~ attr(.x, "correlation")[1, 2])
@@ -327,6 +392,8 @@ icc.lme4 <- function(fit, obj.name) {
 
       # bind to matrix
       tau.01 <- apply(cbind(rho.01, std_), MARGIN = 1, FUN = prod)
+
+      message("Caution! ICC for random-slope-intercept models usually not meaningful. See 'Note' in `?icc`.")
     }
 
     # name values
@@ -351,6 +418,7 @@ icc.lme4 <- function(fit, obj.name) {
     attr(ri.icc, "rho.01") <- rho.01
     attr(ri.icc, "tau.11") <- tau.11
     attr(ri.icc, "sigma_2") <- resid_var
+    attr(ri.icc, "rnd.slope.model") <- any(has_rnd_slope)
 
 
     # finally, save name of fitted model object. May be needed for
@@ -361,8 +429,118 @@ icc.lme4 <- function(fit, obj.name) {
     # return results
     return(ri.icc)
   } else {
-    warning("Function `icc` currently only supports `merMod` (package `lme4`), `glmmTMB` (package `glmmTMB`) or `brmsfit` (package brms) objects.", call. = TRUE)
+    warning("`icc()` does not support this model-object.", call. = TRUE)
   }
+}
+
+
+#' @importFrom purrr map_df map_if
+icc.posterior <- function(fit, obj.name) {
+  if (inherits(fit, "brmsfit") && !requireNamespace("brms", quietly = TRUE))
+    stop("Please install and load package `brms` first.", call. = F)
+
+  # get family
+  fitfam <- stats::family(fit)$family
+
+  # is neg. binomial?
+  is_negbin <-
+    sjmisc::str_contains(
+      fitfam,
+      c("Negative Binomial", "nbinom", "negbinomial"),
+      ignore.case = TRUE,
+      logic = "OR"
+    )
+
+  # is logistic?
+  is_logistic <- fitfam %in% c("bernoulli", "binomial")
+
+  # get random effect variances for each sample of posterior
+  reva <- brms::VarCorr(fit, summary = FALSE)
+
+  # remove "residual__" element from list
+  # and save in separate object
+  reva.resid <- reva[names(reva) == "residual__"]
+  reva <- reva[!(names(reva) == "residual__")]
+
+
+  # retrieve only intercepts
+  vars <- purrr::map(reva, ~ .x$sd[, 1] ^ 2)
+
+  # random intercept-variances, i.e.
+  # between-subject-variance (tau 00)
+  tau.00 <- purrr::map(vars, ~ .x)
+
+  # random slope-variances (tau 11)
+  tau.11 <- purrr::map(reva, ~ .x$cov[, 2, 2])
+
+  # get residual standard deviation sigma
+  sig <- reva.resid[["residual__"]]$sd[, 1]
+
+  # set default, if no residual variance is available
+  if (is.null(sig)) {
+    if (is_logistic)
+      sig <- sqrt((pi ^ 2) / 3)
+    else
+      sig <- 1
+  }
+
+
+  # residual variances, i.e.
+  # within-cluster-variance (sigma^2)
+
+  if (is_logistic) {
+    # for logistic models, we use pi / 3
+    resid_var <- (pi ^ 2) / 3
+  } else {
+    # for linear and poisson models, we have a clear residual variance
+    resid_var <- sig ^ 2
+  }
+
+
+  # total variance, sum of random intercept and residual variances
+  total_var <- apply(as.data.frame(vars), MARGIN = 1, FUN = sum) + resid_var
+
+
+  # check whether we have negative binomial
+
+  if (is_negbin) {
+
+    # for negative binomial models, we also need the intercept...
+    beta <- as.numeric(brms::fixef(fit)[[1]])
+    # ... and the theta value to compute the ICC
+    r <- sig
+
+    # make formula more readable
+
+    numerator <- purrr::map(tau.00, ~ exp(.x) - 1)
+    denominator <- ((exp(total_var) - 1) + (exp(total_var) / r) + exp(-beta - (total_var / 2)))
+
+    ri.icc <- purrr::map(numerator, ~ .x / denominator)
+  } else {
+    # random intercept icc
+    ri.icc <- purrr::map(tau.00, ~ .x / total_var)
+  }
+
+  tau.11 <- purrr::map_if(tau.11, is.null, ~ rep(NA, length(resid_var)))
+
+  names(ri.icc) <- sprintf("icc_%s", names(ri.icc))
+  names(tau.00) <- sprintf("tau.00_%s", names(tau.00))
+  names(tau.11) <- sprintf("tau.11_%s", names(tau.11))
+
+  icc_ <- dplyr::bind_cols(ri.icc, tau.00, tau.11, data.frame(resid_var = resid_var))
+
+  attr(icc_, "family") <- stats::family(fit)$family
+  attr(icc_, "link") <- stats::family(fit)$link
+  attr(icc_, "formula") <- stats::formula(fit)
+  attr(icc_, "model") <- "Bayesian mixed model"
+  attr(icc_, "tau.00") <- tau.00
+  attr(icc_, "tau.11") <- tau.11
+  attr(icc_, "sigma_2") <- resid_var
+
+  class(icc_) <- c("icc.posterior", class(icc_))
+
+  # return results
+  icc_
 }
 
 
@@ -426,7 +604,10 @@ re_var <- function(x) {
   # iterate all attributes and return them as vector
   rv <- c("sigma_2", "tau.00", "tau.11", "tau.01", "rho.01")
 
-  rv_ <- purrr::map(rv, ~ attr(icc(x), .x, exact = TRUE))
+  # compute icc
+  icc_ <- suppressMessages(icc(x))
+
+  rv_ <- purrr::map(rv, ~ attr(icc_, .x, exact = TRUE))
   rn <- purrr::map2(1:length(rv_), rv, ~ sjmisc::trim(paste(names(rv_[[.x]]), .y, sep = "_")))
   rv_ <- purrr::flatten_dbl(rv_)
 
@@ -450,7 +631,7 @@ get_re_var <- function(x, comp = c("tau.00", "tau.01", "tau.11", "rho.01", "sigm
   comp <- match.arg(comp)
 
   # do we have a merMod object? If yes, get ICC and var components
-  if (is_merMod(x) || inherits(x, c("glmmTMB", "brmsfit"))) x <- icc(x)
+  if (is_merMod(x) || inherits(x, c("glmmTMB", "brmsfit"))) x <- suppressMessages(icc(x))
 
   # return results
   attr(x, comp, exact = TRUE)